Update decimal precision, config, and ClickOnce settings
- Change TaxRate and CurrencyRate types from double to decimal for improved financial accuracy in models and worker logic.
- Comment out CreateDownPaymentInvoiceTask() to skip down payment invoice processing.
- Switch API base URL to production in appsettings.json; add separate staging URL.
- Add staging API credentials as comments in APIService.cs.
- Add ClickOnce deployment and manifest signing settings to project file, including temporary key for signing.

+++ SAP-LHDN/Models/Base.cs
@@ -27,7 +27,7 @@ namespace SAP_LHDN.Models
         [JsonProperty("TaxType")]
         public string TaxType { get; set; } = string.Empty;
         [JsonProperty("TaxRate")]
-        public double TaxRate { get; set; }
+        public decimal TaxRate { get; set; }
         [JsonProperty("TaxAmount")]
         public decimal TaxAmount { get; set; }
         [JsonProperty("TaxExemption")]
@@ -92,7 +92,7 @@ namespace SAP_LHDN.Models
         [JsonProperty("Currency")]
         public string Currency { get; set; } = string.Empty;
         [JsonProperty("CurrencyRate")]
-        public double CurrencyRate { get; set; }
+        public decimal CurrencyRate { get; set; }
         [JsonProperty("Terms")]
         public string Terms { get; set; } = string.Empty;
